# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The local daemon's API (ADR 0010): what every surface that is not an agent reads and changes. It serves derived figures only, never prompt text, file contents or transcripts. Every route needs the bearer token from ~/.tokenops/daemon.url.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "TokenOps daemon API",
    "version": "local"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/coach": {
      "get": {
        "operationId": "getCoach",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "compaction": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "client": {
                            "type": "string"
                          },
                          "key": {
                            "type": "string"
                          },
                          "note": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "value": {
                            "type": "integer"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "follow_through": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "channel": {
                            "type": "string"
                          },
                          "followed": {
                            "type": "integer"
                          },
                          "ignored": {
                            "type": "integer"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "open": {
                            "type": "integer"
                          },
                          "power": {
                            "type": "string"
                          },
                          "quiet": {
                            "type": "boolean"
                          },
                          "stood": {
                            "type": "integer"
                          },
                          "undone": {
                            "type": "integer"
                          },
                          "unknown": {
                            "type": "integer"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "hooks": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "client": {
                            "type": "string"
                          },
                          "detail": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "model_policy": {
                      "additionalProperties": false,
                      "properties": {
                        "Allow": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "Deny": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    },
                    "powers": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "configured": {
                            "type": "string"
                          },
                          "describes": {
                            "type": "string"
                          },
                          "effective": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "note": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "preset": {
                      "type": "string"
                    },
                    "verbosity": {
                      "type": "string"
                    },
                    "verbosity_source": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "The coach's dials and each power's autonomy"
      },
      "post": {
        "operationId": "postCoach",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "autonomy": {
                    "type": "string"
                  },
                  "power": {
                    "type": "string"
                  },
                  "preset": {
                    "type": "string"
                  },
                  "rung": {
                    "type": "string"
                  },
                  "verbosity": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "compaction": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "client": {
                            "type": "string"
                          },
                          "key": {
                            "type": "string"
                          },
                          "note": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "value": {
                            "type": "integer"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "follow_through": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "channel": {
                            "type": "string"
                          },
                          "followed": {
                            "type": "integer"
                          },
                          "ignored": {
                            "type": "integer"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "open": {
                            "type": "integer"
                          },
                          "power": {
                            "type": "string"
                          },
                          "quiet": {
                            "type": "boolean"
                          },
                          "stood": {
                            "type": "integer"
                          },
                          "undone": {
                            "type": "integer"
                          },
                          "unknown": {
                            "type": "integer"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "hooks": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "client": {
                            "type": "string"
                          },
                          "detail": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "model_policy": {
                      "additionalProperties": false,
                      "properties": {
                        "Allow": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "Deny": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    },
                    "powers": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "configured": {
                            "type": "string"
                          },
                          "describes": {
                            "type": "string"
                          },
                          "effective": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "note": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "preset": {
                      "type": "string"
                    },
                    "verbosity": {
                      "type": "string"
                    },
                    "verbosity_source": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Apply a coach preset, or change its dials"
      }
    },
...
    "/api/config": {
      "get": {
        "operationId": "getConfig",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "The active configuration, secrets redacted"
      }
    },
    "/api/data-sources": {
      "get": {
        "operationId": "getDataSources",
        "parameters": [
          {
            "description": "RFC 3339 time or a duration such as 24h",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "RFC 3339 time; open when omitted",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "counts": {
                      "properties": {},
                      "type": "object"
                    },
                    "error": {
                      "type": "string"
                    },
                    "freshness_unavailable": {
                      "type": "string"
                    },
                    "hint": {
                      "type": "string"
                    },
                    "sources": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "events_in_window": {
                            "type": "integer"
                          },
                          "last_error": {
                            "type": "string"
                          },
                          "last_error_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "last_event_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "last_poll_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "last_successful_poll_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "never_seen": {
                            "type": "boolean"
                          },
                          "origin_newest_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "reader_stopped": {
                            "type": "boolean"
                          },
                          "severity": {
                            "type": "string"
                          },
                          "silent_for": {
                            "type": "integer"
                          },
                          "state": {
                            "type": "string"
                          },
                          "tag": {
                            "type": "string"
                          },
                          "window": {
                            "type": "integer"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "unhealthy": {
                      "type": "integer"
                    },
                    "window": {
                      "additionalProperties": false,
                      "properties": {
                        "since": {
                          "type": "string"
                        },
                        "until": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Events per source and each reader's health"
      }
    },
...
    "/api/mode": {
      "get": {
        "operationId": "getMode",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "anything_acts": {
                      "type": "boolean"
                    },
                    "authority": {
                      "type": "string"
                    },
                    "budgets": {
                      "type": "integer"
                    },
                    "held_back": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "held_back_note": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string"
                    },
                    "routing_rules": {
                      "type": "integer"
                    },
                    "subsystems": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "configured": {
                            "type": "string"
                          },
                          "describes": {
                            "type": "string"
                          },
                          "effective": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "setting": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Operating mode and what each subsystem may do"
      },
      "post": {
        "operationId": "postMode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "mode": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "config": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Set the operating mode"
      }
    },
...
    "/api/status": {
      "get": {
        "operationId": "getStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "blockers": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "insight": {
                      "additionalProperties": false,
                      "properties": {
                        "level": {
                          "description": "clear when service checks pass, attention when running with reduced coverage, action_required when configuration blockers remain, unavailable when readiness has not been established",
                          "type": "string"
                        },
                        "summary": {
                          "description": "concise interpretation of readiness, blockers, and warnings; detailed evidence remains in sibling fields",
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "next_actions": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "ready": {
                      "type": "boolean"
                    },
                    "state": {
                      "type": "string"
                    },
                    "warnings": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Readiness, blockers, warnings and next actions"
      }
    },
...
    "/api/vendor-usage": {
      "get": {
        "operationId": "getVendorUsage",
        "parameters": [
          {
            "description": "lookback in hours; 24 when omitted",
            "in": "query",
            "name": "window_hours",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "hint": {
                      "type": "string"
                    },
                    "sources": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "always_on": {
                            "type": "boolean"
                          },
                          "config_hint": {
                            "type": "string"
                          },
                          "enabled": {
                            "type": "boolean"
                          },
                          "events_in_window": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "source_tag": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "window_hours": {
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "The request was refused; the body says why"
          },
          "401": {
            "description": "Missing or wrong API token"
          }
        },
        "summary": "Which usage sources are on and producing"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "http://127.0.0.1:7878"
    }
  ]
}